pyproject.toml
[tool.poetry]
package-mode = false

[project]
name = "shop"
version = "0.1.0"
description = ""
authors = [
    { name = "KateySt", email = "kateystetsenko@gmail.com" }
]
requires-python = ">=3.11"
dependencies = ["passlib (>=1.7.4,<2.0.0)", "password-strength (>=0.0.3.post2,<0.0.4)", "bcrypt (>=5.0.0,<6.0.0)", "pre-commit (>=4.5.1,<5.0.0)", "sentry-sdk[fastapi] (>=2.55.0,<3.0.0)", "orjson (>=3.11.7,<4.0.0)", "logtail-python (>=0.3.4,<0.4.0)", "pyjwt (>=2.12.1,<3.0.0)", "redis (>=7.3.0,<8.0.0)"]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 150
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
exclude = "^alembic/"

[[tool.mypy.overrides]]
module = [
    "alembic.*",
    "alembic.versions.*",
    "alembic.env",
]
ignore_errors = true

[dependency-groups]
dev = [
    "types-redis (>=4.6.0.20241004,<5.0.0.0)"
]
